tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/initial.reset.same.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: initial.reset.same</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>initial.reset.same</h1>
<p class="desc">Setting size (not changing the value) resets canvas to transparent black</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="initial.reset.same.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Setting size (not changing the value) resets canvas to transparent black");
_addTest(function(canvas, ctx) {

canvas.width = 100;
ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 50, 50);
_assertPixel(canvas, 20,20, 255,0,0,255, "20,20", "255,0,0,255");
canvas.width = 100;
_assertPixel(canvas, 20,20, 0,0,0,0, "20,20", "0,0,0,0");


});
</script>
